+++ CHANGES.md
@@ -3,7 +3,7 @@
 ## 1.2.0
 
 - #3291 Remove log4j from scala codebase dependency list
-- Upgrade elasticsearch to 6.8.22
+- #3293 Upgrade elasticsearch to 6.8.22
 - Upgrade logback to 1.2.10
 
 ## 1.1.0