+++ CHANGES.md
@@ -8,6 +8,7 @@
 - #3294 Indexer will auto-fixes non-topologically closed Polygons / MultiPolygons
 - Crawler will now wait for 3 seconds (used to be 1 second) before perform trim action to avoid timeout issue
 - Refactor Indexer code to remove the redundant index queue
+- Allow overiding publisher / format indice version number for search API
 
 ## 1.1.0